back-end/requirements.txt
Django==1.10.2
django-cors-headers==1.2.2
djangorestframework==3.4.7
numpy==1.11.2
pandas==0.19.0
python-dateutil==2.5.3
pytz==2016.7
rest-pandas==0.4.1
six==1.10.0
